test(root): setup repo-wide tests with jest
Some of the projects in the repo have been configured inidividually for
tests.

However, we have decided to follow a root-config approach for our tests.
We have decided to follow this approach to benefit from convenience
brought by runnning all tests from a single command as in this lerna
project.