chore(root): update typescripts configurations
As part of cleaning up the project and strengthening consistency in the
in the project, we have decided to rename the root tsconfig file to a more
clear meanigful name that clearly defines the purpose of the file.

In addition to this, we have added another tsconfig file that takes
advantage of the recently discovered 'references' config as by the
typescript team.

There's a couple of benefits for using this 'refereces' config,
especially for monorepos, but we have currently used it in its most
basic form. Overtime, we will expand on this config.