test(kotii-ui/components/controls): correct component names references
As part of cleaning up the project and strengthening consistency in the
in the project, we have decided to rename the root tsconfig file to a more
clear meanigful name that clearly defines the purpose of the file.

Component names were incorrectly referenced in test files. We are correcting
that for controls category of the components with this commit.