fix(kotii-ui): update compnents to use data-testid for tests
At this point in time, about 70% of kotii-ui compnents are wrappers
around Grommet component library. And for the purpose of testing(renders
of components), we currently only want to see if the wrapping components
render successfully.

As a result, we are using react testing library's data-testid attribute
to mark the components we want to test renders for. Looking at the best
practices for testing react with @testing-library, we are doing this as
an escape hatch for now.