fix(kotii-ui): fix all broken components tests
In the first iteration of kotii-ui components, some conponents tests
were not defined, while some had broken implementations. This a known
issue that was scheduled for a fix.

Some had their types improperly annotated. This commit should be able
to fix all of the related issues except for two or
so warnings that require some time. We will dedicate time to these
before our mvp launch.