test(root): create pre-commit hook
We want to run tests before every commit on our source code to ensure
code quality, consistence, and intended functionality. For this, we have
added a pre-commit hook in husky.

To better run our tests with improved performance, we will take advantage
of lerna and nx's cache functinality to cache test results.