test(kotii-utils): update methods to return a value
I think it as a good practice to notify a user about the outcome of an
operation. This is a practice that we have decided to follow in our utils
localStorage methods.

This is especially important in automated tests where we have to test for an
outcome of some action taken.