chore(root/anzii): switch package manager
After a long time of using yarn, we have decided to switch back to npm.
Yarn is a very good package manager that excels in many aspects, however,
we have found its use a little inconvenient in some cases, and one
case that we can point to is local package installation using tarballs.

Trying to install with yarn proved to be a challenging task that
happened to be very quick and convenient with npm. Further research on
tarball installation using yarn showed that the support for this
behavior is not clear with yarn.

Although, this case was the main driving force behind us swtiching to
npm, there are other challenges that we have faced in the past with yarn,
such the perceived inconsistencies in its manintenace.