chore(kotii-scripts): remove node experimental warning
kotii-scripts uses node hooks to hook into a the system's module

resolution mechanism. At some point, an --experimental-loader flag was
require for this. Overtime, an alternative was provided and this flag
resulted in an --experimental-warning on the console for where is used.

This commit removes the initial use of the flag with suggested approach.