chore(root): switch from npm to pnpm
Due to dependecies being not properly installed from .tgz for local testing,
we've made a decision to switch to pnpm from npm.

With this change, we can properly choose which packages to hoist and which
to install on specific packages in our monorepo.

All the files in this commit are associated to this switchup