chore(kotii-scripts): sets project package manager type
This change is made to determine wether the project uses pnpm package manager.

This is due to a special case of pnpm's way of resolving modules.

We had everything running as expected on projects using yarn and npm,
but for pnpm, kotiiJS was not functioning as expected.

The inconsistency in behaviour was due to how pnpm resolves modules.

With this change, we will have an internal check for pnpm to be able to properly cater for
its use case