chore(kotii-scripts): creates universal hook
The universal hook in kotiiJS will be used to mimic React's useEffect
hook's behaviour on the server. Currently, React does not run useEffect
on the server. Nonetheless, knowing useEffect's goal in React in general,
which is providing a hacky way to run async code, we researched to find
a way of emulating the same behaviour on the server as well.

With useUniversalEffect, we can have useEffect-like goal achived
server side, and then hydrate the data on the client.