chore(kotii-scripts): checks if effect id existence
useUniversalEffect identifies effects in the order that they appear
in the source code, and the same set of effects will be executed in that
order on the server. The effects are viewed as organised in a sequencial
list, as a result, a counter variable name currentEffectID is used to
to keep track of the current effect to properly access its data.

Nonetheless, incrementing the counter without reseting it creates data
access issues in the lifetime of the current process's execution as the
currentEffectID will increment with IDs of data and effects that do not
exist.

To prevent this issue, we use a logic that keeps track of the
currentEffectID, if the ID points to a none-existent data identifier in
the effect's data set, we reset the counter so that a follow up
execution of the useUniversalEffect hooks can properly identify correnct
data