chore(kotii-scripts): creates runComponentEffects method
This method will be used to run possible effects for a given
component pre-component rendering. The component will have its required
data readily available at a point of rendering. This also ensures that
the server only renders the component once as opposed to its usual react-ish execution
on the client