chore(kotii-scripts): change kotii land path and babel config
This commits updates the path to app_routes, previously, this file was
generated and stored in kotiiJS root, but it has since been moved to
kotii's kotii-land folder's root.

We also modified the way we manipulate the babel.server.build.json file,
with the modification, we check if the file already has the scoped-styles-plugin set,
if it is, we skip modifying the file content.