chore(kotii-templates): creates assets meta
This commits adds a logic to save png base64 string to a json file in the system.
The Json file will be used as a temp storage for png
imports base64 strings that are generated for every png file imported.

Webpack will use a loader that returns a saved base64 string for every request