chore(kotii-scripts): store renderd css by specifier
At build time, we needed a way to retrieve a css modules specifier's
mapped values. We have implemented a change that enables us to save a
css module's map by the specifier.

Using a specifier to store the module's map makes it possible for us to
retrieve that module's map during build-time.