fix(kotii-scripts): fix a logical reference error
Since the use of the specifier to store a css module's class name map,
we had an error where the saved object of maps was empty.
The error was due to referencing a full module path instead of the
specifier.

The fix was done to correctly store the map object by a specifier.