chore(kotii-scripts): create local file-loader
As the functionality is before this change, it's a little not reliable
how hydration works works between the server and webpack processed
images. The file-loader functionality will enable us a full control in
working with image and related assets.

The file-loader has been created to help us hydrate our resources with
confidence. It will also integrate with other parts of our system with
ease.