chore(kotii-scripts): create imports context
For some of the static file maps that kotiijs generates, an import
specifier is used to store a map. This works fine for nodejs compilation
process, but webpack does this differently.

Webpack generates a context object that contains some information about
the import. The context generated does not include information about the
exact import specifier, and as a result, we cannot reference some of the
resources that are using original import specifiers as their ids in
webpack.

Webpack appends information such as the import's relative url in its
import path information, and to align with this, we have implemented
functionality in this this commit to slighlty mimic webpack's
pathContext.

This change will ensure that we effectively and confidently work with
webpack without changing our import references strategy.