chore(kotii-scripts): use relative path for data maps reference
From the pathContext object that kotiijs has implemented, we use some of
the information it contains to reference certain resources from webpack.
This information referes to the filenames and related data was generated
by kotiijs during nodejs compilation in both resolver/loader hooks.

This is more a syncing technique that ensures that our webpack bundle
matches with our server-compiled resources. This in turn helps with a
much better hydration during render.