chore(kotii-scripts): update assets json object
We replaced file-loader with custom kotiijs file processing logic, but
along with this change came a need to save static assets to a user's static
folder.

Kotiijs's file processing logic was only either returning path strings or
data for the various files it processed.

To be able to save files in addition to returning path strings or data,
this commit brings a change that allows us to save files for returned
path strings while skipping files that return data.