chore(kotii-scripts): add shortenend absolute path for route
Kotiijs generates a route for each page or screen. This route is an
object with properties. The object contains a key that contains an absolute
path to the page file.

The absolute path key's content is okay for development because it can
be updated when kotiijs' project is executed from system to system.

However, this behavior changes in production, and in a case where a
project is exported as a standalone, an issue may rise where the path will
be invalid in another system. This is because kotiijs does not update this
path was it is created at build time.

This change ensures that page path is valid regardless of the environment where
kotiijs runs.