chore(kotii-scripts): resolve userland resources
This change implemennts a logic for resolving resources in what we refer
to as user-land in kotiijs.

We are basically just resolving the dynamic-ity of the resources.

Just as an example, resolving a layout file from the user-land was only
assumed to be coming from a .js file,
but in reality, users can have varying extensions for their layout files
such as .ts, .jsx, .tsx,etc

This commit will ensure that resolving these resources will be done sucessfully
regardless of the type of the javascript file.