docs(root): add webpack docs
This change add webpack documentation. The content of the documentation is a
deep dive into the internals of webpack. It is a step-by-step guide into
how webpack works functionally in every step of its execution.

This change has been added to help provide clear answers into webpack
questions that developers of kotiijs and users alike may have.